pyproject.toml
[tool.poetry]
name = "crypto-env"
version = "0.1.0"
description = ""
authors = ["Arnold <arnold-enrique@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "1.4.53"
pandas = "1.5.3"
psycopg2 = "2.9.9"
requests = "2.32.3"
sqlalchemy-redshift = "0.8.14"
numpy = "1.26.4"
python-dotenv = "1.0.1"
urllib3 = "2.2.2"
tzdata = "2024.1"
typing-extensions = "4.12.2"
pytz = "2024.1"
python-dateutil = "2.9.0.post0"
idna = "3.8"
greenlet = "3.0.3"
charset-normalizer = "3.3.2"
certifi = "2024.8.30"
redshift-connector = "^2.1.3"
awswrangler = "^3.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
